# feat: make hightlighting linear