@matklad 2021-11-27
https://youtu.be/wvEgymUm7cY <- me not understanding what's going on here :-)