@flodiebold 2021-11-28
Even if it doesn't improve performance, you'd expect _some_ kind of change in the instruction counts, wouldn't you? :confused: