@matklad 2021-12-05
I've more carefully re-did laziness checking, and for analysis stats it seems that although *some* reverse maps are indeed not queried, this is not the common case. I've also gave a shot at using more efficient hash map: 

```rust
// better hash map
Database loaded:     593.14ms, 306minstr, 119mb (metadata 388.94ms, 21minstr, 882kb; build 112.76ms, 8732kinstr, -252kb)
  crates: 39, mods: 827, decls: 18716, fns: 13959
Item Collection:     9.66s, 74ginstr, 382mb
  exprs: 384097, ??ty: 387 (0%), ?ty: 285 (0%), !ty: 144
Inference:           43.54s, 340ginstr, 644mb
Total:               53.19s, 414ginstr, 1027mb

// optimization from last week
Database loaded:     596.02ms, 306minstr, 119mb (metadata 388.88ms, 21minstr, 882kb; build 111.12ms, 8743kinstr, -232kb)
  crates: 39, mods: 827, decls: 18716, fns: 13959
Item Collection:     9.57s, 74ginstr, 390mb
  exprs: 384097, ??ty: 387 (0%), ?ty: 285 (0%), !ty: 144
Inference:           42.71s, 340ginstr, 645mb
Total:               52.27s, 414ginstr, 1036mb

// baseline
Database loaded:     604.75ms, 306minstr, 119mb (metadata 395.82ms, 21minstr, 882kb; build 112.13ms, 8811kinstr, -252kb)
  crates: 39, mods: 827, decls: 18716, fns: 13959
Item Collection:     9.81s, 74ginstr, 378mb
  exprs: 384097, ??ty: 387 (0%), ?ty: 285 (0%), !ty: 144
Inference:           42.39s, 340ginstr, 642mb
Total:               52.20s, 414ginstr, 1020mb
```

It still uses more memory, but not as much memory as the original version.  I think it makes sense to ship this -- more memory is unfortunate, but I feel not having O(N^2) paths is more so. 

While working on it, I also got the impression that maybe we *don't* need to have AstIdMap at all? It seems that it should be possible to move ItemTree into hir_expand, and than make `AstIdMap` data just a  part of that.